Move MigrationRunner to its own package
